feat: move signup functions to its own composable
