+++ src/runtime/composables/useBeditaAuth.ts
@@ -1,7 +1,7 @@
 import { useRecaptcha } from '../composables/useRecaptcha';
 import { useUserState } from '../states/user';
-import { computed, type ComputedRef, useRoute } from '#imports';
-import type { SignupBeditaBody, UserAuth } from '../types';
+import { computed, type ComputedRef } from '#imports';
+import type { UserAuth } from '../types';
 import { filterUserDataToStore } from '../utils/user-data-store';
 import { RecaptchaActions } from '../utils/recaptcha-helpers';
 
@@ -33,33 +33,10 @@ export const useBeditaAuth = () => {
     user.value = null;
   };
 
-  const signup = async (data: SignupBeditaBody) => {
-    const recaptcha_token = await executeRecaptcha(RecaptchaActions.SIGNUP);
-    return await $fetch('/api/bedita/signup', {
-      method: 'POST',
-      body: {
-        ...data,
-        recaptcha_token
-      },
-    });
-  };
-
-  const signupActivation = async (uuid?: string) => {
-    const route = useRoute();
-
-    return await useFetch('/api/bedita/signup/activation',{
-      method:'POST',
-      body: { uuid: uuid || route.query?.uuid },
-      server: false,
-    });
-  };
-
   return {
     user,
     isLogged,
     login,
     logout,
-    signup,
-    signupActivation,
   };
 }
